Drop support for running tests in Firefox.
The latest version of Firefox (52.0 as of this commit) requires
selenium 3.x and geckodriver. scala-js-env-selenium depends on
selenium 2.53.0. scala-js/scala-js-env-selenium#62 should fix this.

Additionally, only firefox-esr is available in Debian Jessie (the OS
for the openjdk:8 image) repositories.

Could I install the latest firefox and figure out how to support it
with the current version of scala-js-env-selenium? Maybe.
Could I just install firefox-esr locally and use the existing
version of scala-js-env-selenium locally and in CI? Yes.
Do I care enough? No.